% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{Fire}
\alias{Fire}
\title{Prices of 50,000 round cut diamonds.}
\format{
A data frame with 15 rows and 2 variables:
\describe{
  \item{DISTANCE}{distance, in miles from Fire Station}
  \item{DAMAGE}{fire damage, in thousands of dollars}
  ...
}
}
\usage{
Fire
}
\description{
A dataset containing the distance and damage of fires
}
\keyword{datasets}
