% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mybin.R
\name{mybin}
\alias{mybin}
\title{Title MyBin}
\usage{
mybin(iter = 100, n = 10, p = 0.5)
}
\arguments{
\item{iter}{iterations}

\item{n}{sample size}

\item{p}{probablitity}
}
\value{
a relative frequency bar plot
}
\description{
Title MyBin
}
\examples{
mybin(iter=1, n=5, p=0.3)
}
