% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myboot2.R
\name{myboot2}
\alias{myboot2}
\title{Title My Boot 2}
\usage{
myboot2(iter = 10000, x = NA, fun = "mean", alpha = 0.05, cx = 1.5, ...)
}
\arguments{
\item{iter}{iterations as a vector}

\item{x}{sample values as a vector}

\item{fun}{function that will be used to find interval}

\item{alpha}{vector}

\item{cx}{vector}

\item{...}{any other}
}
\value{
a list of xstat and a histogram with the
bootstrap interval and point estimate
}
\description{
Title My Boot 2
}
\examples{
myboot2(iter=10000,x=c(1,2),fun="mean",alpha=0.01)
}
