% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mymaxlikg.R
\name{mymaxlikg}
\alias{mymaxlikg}
\title{Title mymaxlikg}
\usage{
mymaxlikg(lfun = "logbin2", theta)
}
\arguments{
\item{lfun}{a function}

\item{theta}{a vector}
}
\value{
a graph that estimates the maximum likelihood estimate
}
\description{
Title mymaxlikg
}
\examples{
\dontrun{mymaxlikg(lfun="logbin",theta=seq(0,1,length=10000))}
}
