% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Title Myncurve}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{vector of means of normal curve}

\item{sigma}{vector of standard deviations of normal curve}

\item{a}{vector of observations}
}
\value{
a graph
}
\description{
Title Myncurve
}
\examples{
myncurve(4,3,2)
}
