% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/newfunctionfromabove.R
\name{newfunctionfromabove}
\alias{newfunctionfromabove}
\title{newfunctionfromabove}
\usage{
newfunctionfromabove(n, iter, a = 0, b = 5)
}
\arguments{
\item{n}{vector, sample size}

\item{iter}{vector, iterations}

\item{a}{vector, starting value}

\item{b}{vector, end value}
}
\value{
a histogram of the sums of the uniform distributions
}
\description{
newfunctionfromabove
}
\examples{
newfunctionfromabove(10,10,0,5)
}
